+++ src/main/java/net/soulsweaponry/entity/ai/goal/NightProwlerGoal.java
@@ -23,18 +23,17 @@ import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.math.*;
 import net.minecraft.world.World;
 import net.soulsweaponry.config.BossConfig;
+import net.soulsweaponry.entity.mobs.*;
+import net.soulsweaponry.entity.projectile.MoonlightProjectile;
 import net.soulsweaponry.entity.projectile.NightSkull;
+import net.soulsweaponry.entity.projectile.NoDragWitherSkull;
 import net.soulsweaponry.entity.projectile.noclip.*;
 import net.soulsweaponry.entity.util.BlackflameSnakeUtil;
 import net.soulsweaponry.entity.util.DeathSpiralLogic;
-import net.soulsweaponry.entity.mobs.*;
-import net.soulsweaponry.entity.projectile.MoonlightProjectile;
-import net.soulsweaponry.entity.projectile.noclip.NightsEdge;
-import net.soulsweaponry.entity.projectile.NoDragWitherSkull;
-import net.soulsweaponry.registry.EntityRegistry;
-import net.soulsweaponry.registry.SoundRegistry;
 import net.soulsweaponry.particles.ParticleEvents;
 import net.soulsweaponry.particles.ParticleHandler;
+import net.soulsweaponry.registry.EntityRegistry;
+import net.soulsweaponry.registry.SoundRegistry;
 import net.soulsweaponry.util.WeaponUtil;
 import org.jetbrains.annotations.Nullable;
 
@@ -54,6 +53,7 @@ public class NightProwlerGoal extends MeleeAttackGoal {
     private int bonusDmg;
     private int flipCounter;
     public static final float PORTAL_RADIUS = 16f;
+    private int chargedSkullCounter = 0;
 
     public NightProwlerGoal(NightProwler boss, double speed, boolean pauseWhenMobIdle) {
         super(boss, speed, pauseWhenMobIdle);
@@ -97,6 +97,7 @@ public class NightProwlerGoal extends MeleeAttackGoal {
         this.boss.setParticleState(0);
         this.bonusDmg = 0;
         this.flipCounter = 0;
+        this.chargedSkullCounter = 0;
     }
 
     private void checkAndSetAttack(LivingEntity target) {
@@ -301,6 +302,7 @@ public class NightProwlerGoal extends MeleeAttackGoal {
             if (this.boss.isPhaseTwo()) {
                 this.boss.setFlying(false);
             }
+            this.chargedSkullCounter = 0;
         }
     }
 
@@ -758,22 +760,29 @@ public class NightProwlerGoal extends MeleeAttackGoal {
             this.boss.setParticleState(4);
         }
         if (this.attackStatus >= 50 && this.attackStatus <= 220) {
-            for (Entity entity : this.boss.getWorld().getOtherEntities(this.boss, this.boss.getBoundingBox().expand(35))) {
-                if (this.attackStatus % 6 == 0 && entity instanceof LivingEntity target) {
-                    int radius = (int) PORTAL_RADIUS - 6;
-                    Vec3d spawn = new Vec3d(this.boss.getX() + this.boss.getRandom().nextBetween(-radius, radius),
-                            this.boss.getY() + 8f, this.boss.getZ() + this.boss.getRandom().nextBetween(-radius, radius));
-                    Vec3d vec = new Vec3d(target.getX() - spawn.getX(), target.getEyeY() - spawn.getY(), target.getZ() - spawn.getZ());
-                    NightSkull skull = new NightSkull(EntityRegistry.NIGHT_SKULL, this.boss.getWorld());
-                    skull.setPosition(spawn);
-                    skull.setVelocity(vec.getX(), vec.getY(), vec.getZ(), 1.75f, 1f);
-                    skull.setOwner(this.boss);
-                    this.boss.getWorld().spawnEntity(skull);
-                    if (target.isDead() && target.deathTime < 2) {
-                        this.boss.heal(BossConfig.night_prowler_eclipse_healing);
-                        DeathSpiralEntity spiral = new DeathSpiralEntity(this.boss.getWorld(), target.getPos(), 1f);
-                        spiral.setPosition(target.getPos());
-                        this.boss.getWorld().spawnEntity(spiral);
+            if (this.attackStatus % 6 == 0) {
+                this.chargedSkullCounter++;
+                for (Entity entity : this.boss.getWorld().getOtherEntities(this.boss, this.boss.getBoundingBox().expand(35))) {
+                    if (entity instanceof LivingEntity target) {
+                        boolean chargedSkull = this.chargedSkullCounter >= BossConfig.night_prowler_eclipse_skull_waves_before_charged_skull_wave;
+                        int radius = (int) PORTAL_RADIUS - 6;
+                        Vec3d spawn = new Vec3d(this.boss.getX() + this.boss.getRandom().nextBetween(-radius, radius),
+                                this.boss.getY() + 8f, this.boss.getZ() + this.boss.getRandom().nextBetween(-radius, radius));
+                        Vec3d vec = new Vec3d(target.getX() - spawn.getX(), target.getEyeY() - spawn.getY(), target.getZ() - spawn.getZ());
+                        NightSkull skull = new NightSkull(this.boss.getWorld(), this.boss, this.getModifiedDamage(chargedSkull ? 10f : 14f), chargedSkull, BossConfig.night_prowler_eclipse_skulls_destroy_blocks,
+                            chargedSkull ? BossConfig.night_prowler_eclipse_skulls_explosion_power_charged : BossConfig.night_prowler_eclipse_skulls_explosion_power, 100);
+                        skull.setPosition(spawn);
+                        skull.setVelocity(vec.getX(), vec.getY(), vec.getZ(), 1.75f, 1f);
+                        this.boss.getWorld().spawnEntity(skull);
+                        if (target.isDead() && target.deathTime < 2) {
+                            this.boss.heal(BossConfig.night_prowler_eclipse_healing);
+                            DeathSpiralEntity spiral = new DeathSpiralEntity(this.boss.getWorld(), target.getPos(), 1f);
+                            spiral.setPosition(target.getPos());
+                            this.boss.getWorld().spawnEntity(spiral);
+                        }
+                        if (chargedSkull) {
+                            this.chargedSkullCounter = 0;
+                        }
                     }
                 }
             }